# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TraderMemos API
  version: 1.0.0
  description: |
    Self-hosted trading journal API.

    Authenticate with either:
    - A short-lived session JWT from `POST /api/v1/auth/login` (web app)
    - A personal access token (`tm_pat_…`) from Settings → API or `POST /api/v1/access-tokens`

    Send the secret as `Authorization: Bearer <token>`.
servers:
  - url: /
    description: This API host
...
paths:
...
  /api/v1/setup:
    post:
      tags: [Auth]
      summary: Complete first-user setup
      operationId: completeSetup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupRequest"
      responses:
        "201":
          description: Owner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "409":
          $ref: "#/components/responses/Error"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register (when registration is open)
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "403":
          $ref: "#/components/responses/Error"
...
  /api/v1/accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      operationId: listAccounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Error"
    post:
      tags: [Accounts]
      summary: Create account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccount"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session JWT access token **or** personal access token (`tm_pat_…`).
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
    Credentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10
        totp_code:
          type: string
          description: >-
            Sent on the second leg of sign-in, after a login answered 401 with
            code `totp_required`. A wrong password answers a plain
            `unauthorized` whether or not the account has an authenticator, so
            the response never reveals that it does.
...
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        is_admin:
          type: boolean
...
    SetupRequest:
      allOf:
        - $ref: "#/components/schemas/Credentials"
        - type: object
          properties:
            account:
              $ref: "#/components/schemas/CreateAccount"
...
    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        broker:
          type: string
        account_type:
          type: string
        base_currency:
          type: string
        starting_balance:
          type: number
    CreateAccount:
      type: object
      required: [name]
      properties:
        name:
          type: string
        broker:
          type: string
        account_type:
          type: string
        base_currency:
          type: string
        starting_balance:
          type: number
security:
  - bearerAuth: []